pyproject.toml
[project]
name = "requirement_tracker"
version = "0.1.0"
description = "A multi-agent AI system that automatically transforms informal requirement descriptions into structured requirement documents."
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.10,<3.14"
license = {text = "MIT"}
keywords = ["AI", "requirements", "documentation", "multi-agent", "crewai"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Documentation",
    "Topic :: Scientific/Engineering :: Artificial Intelligence"
]
dependencies = [
    "python-dotenv>=1.0.0,<2.0.0",
    "streamlit>=1.39.0,<2.0.0",
    "langchain>=0.3.0,<1.0.0",
    "langchain-openai>=0.2.0,<1.0.0",
    "litellm>=1.47.0,<2.0.0",
    "langchain-dashscope>=0.1.2,<1.0.0",
    "crewai==1.6.1",
    "crewai-tools==1.6.1",
    "dashscope",
    "langchain-community==0.3.27",
    "atlassian-python-api",  
    "azure-devops"
]

[project.scripts]
requirement_tracker = "src.main:main"

[project.urls]
Homepage = "https://github.com/loootte/req_agent"
Repository = "https://github.com/loootte/req_agent"
Issues = "https://github.com/loootte/req_agent/issues"

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"*" = ["*.md", "*.txt"]

[dependency-groups]
dev = [
"pytest>=7.0,<8.0",
"pytest-mock>=3.0,<4.0",
"pytest-cov>=4.0,<5.0",
"coverage>=7.0,<8.0"
]

[tool.pytest.ini_options]
markers = [
    "e2e: End-to-end tests that require external services and should be excluded from CI pipeline",
    "integration: Integration tests",
    "unit: Unit tests"
]